requirements.txt
Flask==2.2.3
numpy==1.23.5
torch==2.0.1
transformers==4.30.2
diffusers==0.22.1
huggingface-hub==0.16.4
Pillow==9.5.0
python-dotenv==1.0.1
